Added fetch to the reducer to update the store based on fetch data.
here we intially modified our code in way that it fetch the comments, promotions are fetched from the server.
we started our localproject on localhost:3000 and simultaneously json-server on localhost:3001.
Then our project fetches the data of comments, images from the db.json using fetch.
so with that our images, comments and promotions are fetched from json server.