Here we added error handler to Action Creators.
We added error handlers to the ActionCreators in order to handle the errors in response generated by the server.
It includes if the json-server is not active or there is no active internet connection then the fetching of data
from the live server is not possible.

In that situation the error handler is used so that we can understand y we can't fetch the images from the server.